refactor: replaced classes
