+++ niffler-e-2-e-tests/src/test/java/niffler/data/enums/CurrencyValues.java
@@ -1,4 +1,4 @@
-package niffler.model;
+package niffler.data.enums;
 
 public enum CurrencyValues {
     RUB, USD, EUR, KZT
